refactor: try-catch block changed
