+++ src/fs/delete.js
@@ -1,4 +1,4 @@
-import { unlink, promises } from 'fs';
+import { unlink } from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
@@ -7,17 +7,16 @@ const __dirname = path.dirname(__filename);
 
 const remove = async () => {
     const filesName = path.join(__dirname, 'files/fileToRemove copy.txt');
-
     const errorMessage = 'FS operation failed';
 
     try {
-        unlink(filesName, (err) => {
-            if (err) {
+        await unlink(filesName).catch((error) => {
+            if (error.code === 'ENOENT') {
                 throw new Error(errorMessage);
             }
         });
-    } catch (err) {
-        console.error(err);
+    } catch(err) {
+        console.error(err.message);
     }
 };
