+++ src/fs/read.js
@@ -1,4 +1,4 @@
-import { readFile } from 'fs';
+import { readFile } from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
@@ -7,19 +7,18 @@ const __dirname = path.dirname(__filename);
 
 const read = async () => {
     const fileName = path.join(__dirname, 'files/fileToRead.txt');
-    
     const errorMessage = 'FS operation failed';
 
-    readFile(fileName, 'utf8', (err, data) => {
-        if(err) {
-            if (err.code === 'ENOENT') {
+    try {
+        const data = await readFile(fileName, 'utf8').catch((err) => {
+            if(err.code === 'ENOENT') {
                 throw new Error(errorMessage);
             }
-            throw err;
-        }
-
+        });
         console.log(data);
-    });
+    } catch (error) {
+        console.error(error.message);
+    }
 };
 
 await read();
